chore(linkStyles): move shared styled-components helper out of Link/ folder
The `linkStyles` and `StyledLinkProps` helpers are a shared styled-components
`css` template used by DateDetails and exposed publicly (deprecated) via the
package root. Move the file from `src/components/Link/common.ts` to
`src/styles/linkStyles.ts` so the Link component's folder can be migrated
to CSS Modules without the helper file's `import 'styled-components'`
contaminating the migration check (`grep -r 'styled-components'
src/components/Link/`).

Imports updated:
- `src/components/Link/Link.tsx` (last styled-components consumer in Link/)
- `src/components/DateDetails/DateDetails.tsx`
- `src/index.ts` (public re-exports, paths only)

No behavior change. Public exports `linkStyles` and `StyledLinkProps` keep
their names and remain marked `@deprecated`.
