+++ src/components/DateDetails/DateDetails.tsx
@@ -2,7 +2,7 @@ import { styled } from 'styled-components';
 
 import { Popover } from '@/components/Popover';
 import { Text } from '@/components/Text';
-import { linkStyles, StyledLinkProps } from '@/components/Link/common';
+import { linkStyles, StyledLinkProps } from '@/styles/linkStyles';
 import { GridContainer } from '@/components/GridContainer';
 import { Container } from '@/components/Container';
